--- a/torchaudio/functional/functional.py
+++ b/torchaudio/functional/functional.py
@@ -130,6 +130,8 @@ def _get_complex_dtype(real_dtype: torch.dtype):
         return torch.cdouble
     if real_dtype == torch.float:
         return torch.cfloat
+    if real_dtype == torch.half:
+        return torch.complex32
     raise ValueError(f'Unexpected dtype {real_dtype}')