Get rid of pseudo complex type in `F.griffinlim`.

Part of #1337 

`griffinlim` is tested
 - ✅ TorchScript on CPU/CUDA for `float32` and `float64`
 - ✅ `librosa` consistency on CPU
 - ✅ Batch consistency on CPU

**TODO**
 - [x] Autograd test (updated from #1421)
 - [x] Benchmark
 - [x] Remove the hardcoded `cfloat` in the intermediate variable.

## Benchmark

<details><summary>script</summary>

```bash
#!/usr/bin/env bash

OMP_NUM_THREADS=1 numactl --membind 0 --cpubind 0 python -m timeit -s """
import torch
import torchaudio;

batch_size = 32
num_channels = 1
num_freq = 201
num_frames = 400

torch.manual_seed(0);
spec = torch.randn(batch_size, num_channels, num_freq, num_frames, dtype=torch.float32, device='cpu');
griffinlim = torchaudio.transforms.GriffinLim()
""" """
griffinlim(spec)
"""
```

</details>

| device |           master (8d2eeb1)           |                This PR                |
|:------:|:------------------------------------:|:-------------------------------------:|
|    CPU | 1 loop, best of 5: 6.48 sec per loop | 1 loop, best of 5: 3.57 sec per loop  |
|    GPU | 1 loop, best of 5: 128 msec per loop | 1 loop, best of 5: 93.4 msec per loop |