@mthrok 2021-04-06
@anjali411 

The autograd tests are failing for the case `rand_init=True`. However, it does not fail on my local env with PyTorch nightly from March 29th.

```
[conda] pytorch                   1.9.0.dev20210329 py3.8_cuda10.1_cudnn7.6.3_0    pytorch-nightly
```

Failing test log: https://app.circleci.com/pipelines/github/pytorch/audio/5665/workflows/746a53f5-6f18-4f2e-b717-869bf342eaf4/jobs/194528

I am wondering if this is a regression. What do you think? The following is the corresponding code;

https://github.com/mthrok/audio/blob/17f61e73f8f4b0f314bb7356aa3a2222c4038347/torchaudio/functional/functional.py#L193-L201