@mthrok 2021-04-07
> > @anjali411
> > The autograd tests are failing for the case `rand_init=True`.
> 
> @mthrok
> 
> Hmm, I think it's an expected behavior, cuz numerical method will (I'm not sure though) forward the function multiple times, and with `rand_init=True`, each forward pass will start with different initial phases, result in different output.
> 
> For functions like it which have random behavior internally, maybe a `.backward()` at the end of test is enough, to make sure that gradients can be propogated without problems.

Hi @yoyololicon 

Thanks for the insight. You are right. I added a wrapper that sets the random seed and the test passes now.
(I do not know why I thought it was working on my local env...)